recovery script that works with the new account recovery feature
